add vars_plugins directory
